augment chart svg selector for light theme (#1400)
by default docusaurus uses data URI to display SVGs. However, it turns
out that SVGs that are bigger, docusaurus uses a regular URL and not a
data URI.

for light theme we have a CSS selector, which targets railroad diagram
SVGs. The targeting is not ideal but works well enough for our use case.

This PR augments the selector to also include SVGs which are not
included through data URI but through regular URL