Add questdb.exe install to binaries.md (#1414)
* Add questdb.exe install to binaries.md

Just now trying to "hello world" questdb, needed to run the `install` command first before getting the start command to work.  Should help the next person trying to get it to run in windows for the first time.  I can't speak for the linux case, but I assume it is a similar process.

* Update docs/get-started/binaries.md

---------

Co-authored-by: Amy Wang <107401731+amyshwang@users.noreply.github.com>